requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
PyPDF2==3.0.1
pyzbar==0.1.9
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
